Created writer base class
